Extract shared async-test helpers (pure moves)
- Move FakeClock and advance_clock_on_sleep to tests/helpers/clock.py; they were
  duplicated between the client and rate-limiting suites and would drift. Both
  suites now import from there; the rate-limiting slept fixture keeps its local
  recording behavior on top of the shared clock.
- Point the download tests (token_not_leaked, signed_url_error, zip_slip) at the
  existing patch_signed_download helper instead of copy-pasting its monkeypatch block.
- Strip explicit @pytest.mark.asyncio from the check-run and download sections;
  asyncio auto mode is the convention.
